+++ ethereum/spec/src/main/java/tech/pegasys/teku/spec/datastructures/networking/libp2p/rpc/versions/eip4844/BlobsSidecar.java
@@ -46,8 +46,7 @@ public class BlobsSidecar
           namedSchema("beacon_block_root", SszPrimitiveSchemas.BYTES32_SCHEMA),
           namedSchema("beacon_block_slot", SszPrimitiveSchemas.UINT64_SCHEMA),
           namedSchema(
-              FIELD_BLOBS,
-              SszListSchema.create(blobSchema, specConfig.getMaxBlobsPerBlock().longValue())),
+              FIELD_BLOBS, SszListSchema.create(blobSchema, specConfig.getMaxBlobsPerBlock())),
           namedSchema("kzg_aggregated_proof", SszKZGProofSchema.INSTANCE));
     }
